Change: Degree now true degree rather than upper bound

+++ src/circuit/circuit.rs
@@ -732,10 +732,10 @@ impl CircuitSeq {
     
         for &[a, b, c] in gates {
             // a' = a XOR (b AND NOT c)
-            let not_c = poly_not(polys[c].clone());
-            let term = poly_and(&polys[b], &not_c);
-            let new_a = poly_xor(polys[a].clone(), term);
-            polys[a] = new_a;
+            let not_c = poly_not(polys[c as usize].clone());
+            let term = poly_and(&polys[b as usize], &not_c);
+            let new_a = poly_xor(polys[a as usize].clone(), term);
+            polys[a as usize] = new_a;
         }
     
         // XOR each wire with its initial value x_i so unchanged wires become 0
